+++ src/app/components/button/button.pug
@@ -1,0 +1,1 @@
+button #{text}
