+++ src/app/components/button/button.scss
@@ -1,0 +1,6 @@
+button {
+    padding: 10px 0;
+    min-width: 130px;
+    background: #f0f0f0;
+    border-radius: 2px;
+}
