+++ src/app/features/home/home.feature.js
@@ -1,26 +1,16 @@
-import Router from '../../core/routes';
-import { Header } from '../../components';
+import { Header, Button } from '../../components';
 import template from './home.template.pug';
 import './home.style.scss';
 import jsImage from '../../../assets/js-logo.png';
 
 export default class Home {
-  constructor() {
-    this.title = 'Home page';
-  }
-
-  click() {
-    let home = document.querySelector('app-home');
-    let btn = home.querySelector('button');
-
-    btn.addEventListener('click', () => {
-      new Router().go('/contacts');
-    });
+  constructor(selector) {
+    this.homeEl = document.querySelector(selector);
   }
 
   load() {
-    document.querySelector('#root').innerHTML = template({ title: this.title, jsImage });
-    new Header().load();
-    this.click();
+    this.homeEl.innerHTML = template({ title: this.title, jsImage });
+    new Header('Header component').load();
+    new Button('#button1').load();
   }
 }
