+++ src/app/features/home/home.template.pug
@@ -2,4 +2,4 @@ app-home
     app-header
     h1 #{this.title}
     img(src=jsImage)
-    button Contacts
+    app-button(id="button1", data-text="Contacts", data-href="/contacts")
