--- src/styles/common.scss
@@ -1,6 +1,0 @@
-button {
-    padding: 10px 0;
-    min-width: 130px;
-    background: #f0f0f0;
-    border-radius: 2px;
-}
